Feature case summary api (#626)
* added summary end point

* added method in service and updated query and rowmapper

* updated method in case summary

* completed case summary end point

* added test cases for case summary end point

* addressed code rabbit comments

* updated test case

* exception handle in repository

* changed limit offset to integer

* updated test case for limit offset integer

---------

+++ backend/case/src/main/java/org/pucar/dristi/repository/querybuilder/CaseQueryBuilder.java
@@ -437,10 +437,10 @@ public class CaseQueryBuilder {
     }
     public String addPaginationQuery(String query, List<Object> preparedStatementList, Pagination pagination, List<Integer> preparedStmtArgList) {
         preparedStatementList.add(pagination.getLimit());
-        preparedStmtArgList.add(Types.DOUBLE);
+        preparedStmtArgList.add(Types.INTEGER);
 
         preparedStatementList.add(pagination.getOffSet());
-        preparedStmtArgList.add(Types.DOUBLE);
+        preparedStmtArgList.add(Types.INTEGER);
         return query + " LIMIT ? OFFSET ?";
 
     }
