+++ backend/case/src/main/java/org/pucar/dristi/web/models/Pagination.java
@@ -25,11 +25,11 @@ public class Pagination {
 	@JsonProperty("limit")
 
 	@DecimalMax("100")
-	private Double limit = 10d;
+	private Integer limit = 10;
 
 	@JsonProperty("offSet")
 
-	private Double offSet = 0d;
+	private Integer offSet = 0;
 
 	@JsonProperty("totalCount")
